+++ __tests__/check-tsconfig.spec.ts
@@ -1,5 +1,7 @@
-import { checkTsConfig } from "./check-tsconfig";
 import * as ts from "typescript";
+
+import { checkTsConfig } from "../src/check-tsconfig";
+
 test("checkTsConfig", () => {
 	expect(() =>
 		checkTsConfig({
@@ -10,6 +12,7 @@ test("checkTsConfig", () => {
 	).toThrow(
 		`Incompatible tsconfig option. Module resolves to 'None'. This is incompatible with rollup, please use 'module: "ES2015"' or 'module: "ESNext"'.`,
 	);
+
 	expect(
 		checkTsConfig({
 			fileNames: [],
@@ -17,6 +20,7 @@ test("checkTsConfig", () => {
 			options: { module: ts.ModuleKind.ES2015 },
 		}),
 	).toBeFalsy();
+
 	expect(
 		checkTsConfig({
 			fileNames: [],
